finished registration step 2
