+++ client/app/register/register.module.css
@@ -106,6 +106,7 @@
 .quantity, .unit, .ingredient{
     display: flex;
     flex-direction: column;
+    color: #506264;
 }
 
 .quantity {
@@ -118,4 +119,66 @@
 
 .ingredient {
     width: 45%;
+}
+
+.ingredientList {
+    margin-top: 2rem;
+    padding: 2rem;
+    width: 60%;
+    height: 325px;
+    border-radius: 10px;
+    background-color: rgba(236, 228, 227, .6);
+    color: #506264;
+    overflow-y: auto;  
+    position: relative; 
+}
+
+.scrollableContainer {
+    display: flex;          
+    flex-wrap: wrap;          
+    max-height: 100%;        
+    align-items: flex-start;          
+    justify-content: space-between;  
+}
+
+.ingredientItem {
+    font-size: 19px;
+    padding-bottom: 1rem;
+    display: flex;
+    align-items: center;
+    justify-content: space-between;
+    width: 45%;
+}
+
+.removeButton {
+    color: #506264;
+    font-size: 20px;
+    margin-left: 7px;
+    &:hover {
+        color:#EC4A27;
+        cursor: pointer;
+    }
+}
+
+.navButtons {
+    display: flex;
+    justify-content: space-between;
+    width: 65%;
+}
+
+.navButton {
+    background-color: #EC4A27;
+    font-size: 1.5rem;
+    font-weight: 600;
+    color: white;
+    width: 20%;
+    text-align: center;
+    border-radius: 10px;
+    padding: .7rem 0;
+    margin-top: 2.5rem;
+
+    &:hover {
+        background-color: rgb(233, 69, 15);
+        cursor: pointer;
+    }
 }
